Redefined token types as enum

+++ EDAnyCharacterLexRuleTest.m
@@ -23,7 +23,7 @@
 	NSUInteger len = source.length;
 	for (i = 0; i < len; ++i) {
 		EDLexicalToken *tok = [rule lexInString:source range:NSMakeRange(i, len - i)];
-		GHAssertEquals((NSUInteger) EDUnmatchedCharacterToken, tok.type, @"Type should be EDUnmatchedCharacterToken");
+		GHAssertEquals(EDUnmatchedCharacterToken, tok.type, @"Type should be EDUnmatchedCharacterToken");
 		GHAssertEquals((NSUInteger) i, tok.range.location, @"Location should be start of range");
 		GHAssertEquals((NSUInteger) 1, tok.range.length, @"Length should be 1");
 	}
@@ -34,7 +34,7 @@
 	NSString *source = @" ";
 	
 	EDLexicalToken *tok = [rule lexInString:source range:NSMakeRange(0, 1)];
-	GHAssertEquals((NSUInteger) EDWhitespaceToken, tok.type, @"Type should be EDWhitespaceToken");
+	GHAssertEquals(EDWhitespaceToken, tok.type, @"Type should be EDWhitespaceToken");
 }
 
 @end
