+++ EDDelimitedStringLexRuleTest.m
@@ -27,7 +27,7 @@
 	NSString *source = @"a \"string\" b";
 	EDLexicalToken *tok = [rule lexInString:source range:NSMakeRange(2, source.length - 2)];
 	
-	GHAssertEquals((NSUInteger) EDString1Token, tok.type, @"Type should be EDString1Token");
+	GHAssertEquals(EDString1Token, tok.type, @"Type should be EDString1Token");
 	GHAssertEquals((NSUInteger) 2, tok.range.location, @"String should be found at offset 2");
 	GHAssertEquals((NSUInteger) 8, tok.range.length, @"String should have length 8");
 }
@@ -37,7 +37,7 @@
 	NSString *source = @"a \"string 1\" \"string 2\" b";
 	EDLexicalToken *tok = [rule lexInString:source range:NSMakeRange(2, source.length - 2)];
 	
-	GHAssertEquals((NSUInteger) EDString1Token, tok.type, @"Type should be EDString1Token");
+	GHAssertEquals(EDString1Token, tok.type, @"Type should be EDString1Token");
 	GHAssertEquals((NSUInteger) 2, tok.range.location, @"String should be found at offset 2");
 	GHAssertEquals((NSUInteger) 10, tok.range.length, @"String should have length 10");
 }
@@ -47,7 +47,7 @@
 	NSString *source = @"a \"string \\\"escaped\\\"\" b";
 	EDLexicalToken *tok = [rule lexInString:source range:NSMakeRange(2, source.length - 2)];
 	
-	GHAssertEquals((NSUInteger) EDString1Token, tok.type, @"Type should be EDString1Token");
+	GHAssertEquals(EDString1Token, tok.type, @"Type should be EDString1Token");
 	GHAssertEquals((NSUInteger) 2, tok.range.location, @"String should be found at offset 2");
 	GHAssertEquals((NSUInteger) 20, tok.range.length, @"String should have length 20");
 }
@@ -57,7 +57,7 @@
 	NSString *source = @"a \"string \\\\\"escaped\\\"\" b";
 	EDLexicalToken *tok = [rule lexInString:source range:NSMakeRange(2, source.length - 2)];
 	
-	GHAssertEquals((NSUInteger) EDString1Token, tok.type, @"Type should be EDString1Token");
+	GHAssertEquals(EDString1Token, tok.type, @"Type should be EDString1Token");
 	GHAssertEquals((NSUInteger) 2, tok.range.location, @"String should be found at offset 2");
 	GHAssertEquals((NSUInteger) 11, tok.range.length, @"String should have length 11");
 }
@@ -67,7 +67,7 @@
 	NSString *source = @"a \"unterminated string";
 	EDLexicalToken *tok = [rule lexInString:source range:NSMakeRange(2, source.length - 2)];
 	
-	GHAssertEquals((NSUInteger) EDString1Token, tok.type, @"Type should be EDString1Token");
+	GHAssertEquals(EDString1Token, tok.type, @"Type should be EDString1Token");
 	GHAssertEquals((NSUInteger) 2, tok.range.location, @"String should be found at offset 2");
 	GHAssertEquals((NSUInteger) source.length - 2, tok.range.length, @"String should consume all of the available range");
 }
@@ -77,7 +77,7 @@
 	NSString *source = @"a \"unterminated string\\";
 	EDLexicalToken *tok = [rule lexInString:source range:NSMakeRange(2, source.length - 2)];
 	
-	GHAssertEquals((NSUInteger) EDString1Token, tok.type, @"Type should be EDString1Token");
+	GHAssertEquals(EDString1Token, tok.type, @"Type should be EDString1Token");
 	GHAssertEquals((NSUInteger) 2, tok.range.location, @"String should be found at offset 2");
 	GHAssertEquals((NSUInteger) source.length - 2, tok.range.length, @"String should consume all of the available range");
 }
@@ -87,7 +87,7 @@
 	NSString *source = @"a \"unterminated string\\\" foo\" bar";
 	EDLexicalToken *tok = [rule lexInString:source range:NSMakeRange(2, source.length - 2)];
 	
-	GHAssertEquals((NSUInteger) EDString1Token, tok.type, @"Type should be EDString1Token");
+	GHAssertEquals(EDString1Token, tok.type, @"Type should be EDString1Token");
 	GHAssertEquals((NSUInteger) 2, tok.range.location, @"String should be found at offset 2");
 	GHAssertEquals((NSUInteger) 22, tok.range.length, @"String should stop at first delimiting '\"'");
 }
@@ -97,7 +97,7 @@
 	NSString *source = @"a \"unterminated string aaa\" fooaaaaaa\" bar";
 	EDLexicalToken *tok = [rule lexInString:source range:NSMakeRange(2, source.length - 2)];
 	
-	GHAssertEquals((NSUInteger) EDString1Token, tok.type, @"Type should be EDString1Token");
+	GHAssertEquals(EDString1Token, tok.type, @"Type should be EDString1Token");
 	GHAssertEquals((NSUInteger) 2, tok.range.location, @"String should be found at offset 2");
 	GHAssertEquals((NSUInteger) 36, tok.range.length, @"String should treat sequences 'aaa' like '\\\'");
 }
